improved api and utils structure
